removed templates level and using page level
