--- src/components/templates/CartTemplate.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import Cart from 'components/organisms/Cart';
-
-const CartTemplate = () => {
-  return (
-    <Cart />
-  )
-}
-
-export default CartTemplate
