--- src/pages/MainPage.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import ProductsListTemplate from '../components/templates/ProductsListTemplate';
-
-const MainPage = () => {
-    return (
-        <ProductsListTemplate />
-    );
-};
-
-  export default MainPage
